Python script for Google spreadsheets
This is the current version for the python script that downloads and converts Google spreadsheets into JSON schemas. If they are valid, they are uploaded to MongoDB with database as 'samples' and collection as 'test'.

Currently cannot handle multiple comma separated components. Comma separated components are all treated as one string and not as a list.