process_mailbox and clear_spreadsheet
The validation code can now check the email of nslsii.provenance and check if there is an email notification for new entries to the Google Form. If such email exists, the email is archived and the validation is run on the spreadsheet. If the doc is valid and successfully uploaded to MongoDB, the spreadsheet is cleared of entered values (leaving only key names).